requirements.txt
asgiref==3.5.2
autopep8==1.6.0
camelot-py==0.10.1
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
datefinder==0.7.1
Deprecated==1.2.13
distro==1.7.0
dj-database-url==0.5.0
Django==4.0.5
django-extensions==3.1.5
et-xmlfile==1.1.0
gcloud==0.18.3
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
numpy==1.22.4
oauth2client==4.1.3
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
pandas==1.4.2
pdfminer.six==20220524
pdfplumber==0.7.1
pdfrw==0.4
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.9
PyPDF2==2.3.1
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jwt==3.3.2
pytz==2022.1
regex==2022.6.2
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tabula-py==2.4.0
tabulate==0.8.10
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
Wand==0.6.7
whitenoise==6.2.0
wrapt==1.14.1
